refactor waste complexity hotspots
