+++ packages/waste-management-runtime/src/runtime-types.ts
@@ -1,0 +1,71 @@
+export const createProgress = (input: {
+  readonly completedSteps: number;
+  readonly totalSteps: number;
+  readonly currentPhase: string;
+  readonly currentStepKey: string;
+  readonly details?: Readonly<Record<string, unknown>>;
+}) => ({
+  completedSteps: input.completedSteps,
+  totalSteps: input.totalSteps,
+  currentPhase: input.currentPhase,
+  currentStepKey: input.currentStepKey,
+  details: input.details,
+  lastUpdatedAt: new Date().toISOString(),
+});
+
+export type WasteManagementJobProgress = ReturnType<typeof createProgress>;
+
+export type WasteManagementOperationRuntime = {
+  readonly initializeDataSource: (
+    instanceId: string,
+    payload: import('@sva/plugin-sdk').WasteManagementInitializeJobInput
+  ) => Promise<{
+    readonly durationMs: number;
+    readonly details: Record<string, unknown>;
+  }>;
+  readonly applyMigrations: (
+    instanceId: string,
+    payload: import('@sva/plugin-sdk').WasteManagementApplyMigrationsJobInput
+  ) => Promise<{
+    readonly durationMs: number;
+    readonly details: Record<string, unknown>;
+  }>;
+  readonly importData: (
+    instanceId: string,
+    payload: import('@sva/plugin-sdk').WasteManagementImportJobInput,
+    progressReporter?: {
+      readonly reportProgress: (progress: WasteManagementJobProgress) => Promise<void> | void;
+    }
+  ) => Promise<{
+    readonly durationMs: number;
+    readonly details: Record<string, unknown>;
+  }>;
+  readonly seedData: (
+    instanceId: string,
+    payload: import('@sva/plugin-sdk').WasteManagementSeedJobInput
+  ) => Promise<{
+    readonly durationMs: number;
+    readonly details: Record<string, unknown>;
+  }>;
+  readonly resetData: (
+    instanceId: string,
+    payload: import('@sva/plugin-sdk').WasteManagementResetJobInput
+  ) => Promise<{
+    readonly durationMs: number;
+    readonly details: Record<string, unknown>;
+  }>;
+  readonly syncMainserver: (
+    instanceId: string,
+    payload: import('@sva/plugin-sdk').WasteManagementSyncMainserverJobInput
+  ) => Promise<{
+    readonly durationMs: number;
+    readonly details: Record<string, unknown>;
+  }>;
+  readonly syncWasteTypes: (
+    instanceId: string,
+    payload: import('@sva/plugin-sdk').WasteManagementSyncWasteTypesJobInput
+  ) => Promise<{
+    readonly durationMs: number;
+    readonly details: Record<string, unknown>;
+  }>;
+};
